Remove empty "test"
I'm generally opposed to silencing a warning by adding a dummy block that expressly does *not* do what the warning recommends. If there's no way to test this, then the warning is correct: there's no test. If there is a way to test this, then adding "true" is not a good test.

# Conflicts:
#	Formula/ssh-askpass.rb