`brew services` requires a specific name in label
Not sure if I makes this to complicated but I don't like having duplicate files
in diffrent repos and don't want to take the Homebrew namespace to an external
repo.

Trying to fix https://github.com/theseal/ssh-askpass/issues/36 and it seems according to
https://github.com/Homebrew/homebrew-services/issues/376
that a service in brew needs to have the same name in label as the file name.